chore: refactor AssignTrainers into components
